# skills/openapi-actions.yaml
openapi: 3.1.0
info:
  title: JobPosting API
  version: 1.0.0
  description: Local JobPosting REST API for agent skills and optional ChatGPT Actions.
servers:
  - url: http://127.0.0.1:8000/api
paths:
  /health:
    get:
      operationId: healthCheck
      summary: Health check
      responses:
        "200":
          description: OK
  /jobs:
    get:
      operationId: listJobs
      summary: List scraped jobs
      responses:
        "200":
          description: Job list
    post:
      operationId: createJob
      summary: Create a job manually
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobCreate"
      responses:
        "201":
          description: Created
  /jobs/{job_id}:
    get:
      operationId: getJob
      summary: Get job by ID
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job
    patch:
      operationId: updateJob
      summary: Update job fields
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated
  /sources:
    get:
      operationId: listSources
      summary: List job sources
      responses:
        "200":
          description: Source list
    post:
      operationId: createSource
      summary: Add a source
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SourceCreate"
      responses:
        "201":
          description: Created
  /sources/{source_id}/scrape:
    post:
      operationId: scrapeSource
      summary: Scrape jobs from a source
      parameters:
        - name: source_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scrape results
  /templates:
    get:
      operationId: listTemplates
      summary: List channel templates
      responses:
        "200":
          description: Template list
  /drafts:
    get:
      operationId: listDrafts
      summary: List drafts
      parameters:
        - name: job_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Draft list
  /drafts/generate:
    post:
      operationId: generateDrafts
      summary: Generate drafts from Jinja templates (local, no LLM)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateDraftsRequest"
      responses:
        "200":
          description: Generated drafts
  /drafts/{draft_id}:
    get:
      operationId: getDraft
      summary: Get draft by ID
      parameters:
        - name: draft_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Draft
    patch:
      operationId: updateDraft
      summary: Update draft content (agent polish save)
      parameters:
        - name: draft_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DraftUpdate"
      responses:
        "200":
          description: Updated
  /drafts/{draft_id}/prompt-pack:
    post:
      operationId: buildPromptPack
      summary: Build copy-paste prompt for external AI
      parameters:
        - name: draft_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromptPackRequest"
      responses:
        "200":
          description: Prompt pack
  /drafts/{draft_id}/import:
    post:
      operationId: importAiResult
      summary: Import polished copy from external AI
      parameters:
        - name: draft_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportResultRequest"
      responses:
        "200":
          description: Updated draft
  /drafts/{draft_id}/export:
    get:
      operationId: exportDraft
      summary: Export draft for copy/download
      parameters:
        - name: draft_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Export payload
components:
  schemas:
    SourceCreate:
      type: object
      required: [name, kind, base_url]
      properties:
        name: { type: string }
        kind: { type: string, enum: [greenhouse, lever, ashby, workday, custom] }
        base_url: { type: string }
        enabled: { type: boolean, default: true }
    JobCreate:
      type: object
      properties:
        title: { type: string }
        company: { type: string }
        location: { type: string }
        description_text: { type: string }
        apply_url: { type: string }
        skills:
          type: array
          items: { type: string }
    JobUpdate:
      type: object
      properties:
        title: { type: string }
        company: { type: string }
        location: { type: string }
        description_text: { type: string }
        needs_manual_fill: { type: boolean }
    GenerateDraftsRequest:
      type: object
      required: [job_id, template_ids]
      properties:
        job_id: { type: integer }
        template_ids:
          type: array
          items: { type: integer }
    DraftUpdate:
      type: object
      properties:
        content: { type: string }
        status: { type: string }
    PromptPackRequest:
      type: object
      properties:
        requirement: { type: string }
    ImportResultRequest:
      type: object
      required: [content]
      properties:
        content: { type: string }
        requirement: { type: string }